API/App/requirements.txt
tensorflow>=2.16 
fastapi>=0.110,<0.120
uvicorn>=0.29,<0.30
numpy>=1.22,<1.26
pillow>=9.0,<11.0
python-multipart>=0.0.6